(#3922) - move more stuff out of utils
Feels cleaner when everything is separate files,
rather than having them all inside one big `utils.js`.
