+++ lib/adapter.js
@@ -7,6 +7,8 @@ var EventEmitter = require('events').EventEmitter;
 var upsert = require('./deps/upsert');
 var Changes = require('./changes');
 var Promise = utils.Promise;
+var isDeleted = require('./deps/docs/is-deleted');
+var isLocalId = require('./deps/docs/is-local-id');
 
 /*
  * A generic pouch adapter
@@ -202,7 +204,7 @@ AbstractPouchDB.prototype.put =
   if (error) {
     return callback(error);
   }
-  if (utils.isLocalId(doc._id) && typeof this._putLocal === 'function') {
+  if (isLocalId(doc._id) && typeof this._putLocal === 'function') {
     if (doc._deleted) {
       return this._removeLocal(doc, callback);
     } else {
@@ -305,7 +307,7 @@ AbstractPouchDB.prototype.remove =
   opts.was_delete = true;
   var newDoc = {_id: doc._id, _rev: (doc._rev || opts.rev)};
   newDoc._deleted = true;
-  if (utils.isLocalId(newDoc._id) && typeof this._removeLocal === 'function') {
+  if (isLocalId(newDoc._id) && typeof this._removeLocal === 'function') {
     return this._removeLocal(doc, callback);
   }
   this.bulkDocs({docs: [newDoc]}, opts, yankError(callback));
@@ -468,7 +470,7 @@ AbstractPouchDB.prototype.get =
   if (typeof id !== 'string') {
     return callback(errors.error(errors.INVALID_ID));
   }
-  if (utils.isLocalId(id) && typeof this._getLocal === 'function') {
+  if (isLocalId(id) && typeof this._getLocal === 'function') {
     return this._getLocal(id, callback);
   }
   var leaves = [], self = this;
@@ -546,7 +548,7 @@ AbstractPouchDB.prototype.get =
       }
     }
 
-    if (utils.isDeleted(metadata, doc._rev)) {
+    if (isDeleted(metadata, doc._rev)) {
       doc._deleted = true;
     }
