+++ lib/adapters/idb/idb-bulk-docs.js
@@ -2,6 +2,10 @@
 
 var utils = require('../../utils');
 var errors = require('../../deps/errors');
+var preprocessAttachments =
+  require('../../deps/docs/preprocess-attachments');
+var processDocs = require('../../deps/docs/process-docs');
+var isLocalId = require('../../deps/docs/is-local-id');
 var idbUtils = require('./idb-utils');
 var idbConstants = require('./idb-constants');
 
@@ -30,7 +34,7 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
 
   for (var i = 0, len = docInfos.length; i < len; i++) {
     var doc = docInfos[i];
-    if (doc._id && utils.isLocalId(doc._id)) {
+    if (doc._id && isLocalId(doc._id)) {
       continue;
     }
     doc = docInfos[i] = utils.parseDoc(doc, opts.new_edits);
@@ -48,7 +52,7 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
   var preconditionErrored = false;
   var blobType = api._meta.blobSupport ? 'blob' : 'base64';
 
-  utils.preprocessAttachments(docInfos, blobType, function (err) {
+  preprocessAttachments(docInfos, blobType, function (err) {
     if (err) {
       return callback(err);
     }
@@ -84,10 +88,9 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
     });
   }
 
-  function processDocs() {
+  function idbProcessDocs() {
 
-    utils.processDocs(docInfos, api, fetchedDocs, txn, results,
-      writeDoc, opts);
+    processDocs(docInfos, api, fetchedDocs, txn, results, writeDoc, opts);
   }
 
   function fetchExistingDocs() {
@@ -100,7 +103,7 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
 
     function checkDone() {
       if (++numFetched === docInfos.length) {
-        processDocs();
+        idbProcessDocs();
       }
     }
 
@@ -115,7 +118,7 @@ function idbBulkDocs(req, opts, api, idb, Changes, callback) {
 
     for (var i = 0, len = docInfos.length; i < len; i++) {
       var docInfo = docInfos[i];
-      if (docInfo._id && utils.isLocalId(docInfo._id)) {
+      if (docInfo._id && isLocalId(docInfo._id)) {
         checkDone(); // skip local docs
         continue;
       }
