+++ lib/adapters/idb/idb.js
@@ -2,6 +2,8 @@
 
 var utils = require('../../utils');
 var merge = require('../../merge');
+var isDeleted = require('../../deps/docs/is-deleted');
+var isLocalId = require('../../deps/docs/is-local-id');
 var errors = require('../../deps/errors');
 var idbUtils = require('./idb-utils');
 var idbConstants = require('./idb-constants');
@@ -85,7 +87,7 @@ function init(api, opts, callback) {
       var cursor = event.target.result;
       if (cursor) {
         var metadata = cursor.value;
-        var deleted = utils.isDeleted(metadata);
+        var deleted = isDeleted(metadata);
         metadata.deletedOrLocal = deleted ? "1" : "0";
         docStore.put(metadata);
         cursor.continue();
@@ -113,7 +115,7 @@ function init(api, opts, callback) {
       if (cursor) {
         var metadata = cursor.value;
         var docId = metadata.id;
-        var local = utils.isLocalId(docId);
+        var local = isLocalId(docId);
         var rev = merge.winningRev(metadata);
         if (local) {
           var docIdRev = docId + "::" + rev;
@@ -317,7 +319,7 @@ function init(api, opts, callback) {
         err = errors.error(errors.MISSING_DOC, 'missing');
         return finish();
       }
-      if (utils.isDeleted(metadata) && !opts.rev) {
+      if (isDeleted(metadata) && !opts.rev) {
         err = errors.error(errors.MISSING_DOC, "deleted");
         return finish();
       }
