+++ lib/adapters/leveldb/leveldb.js
@@ -18,6 +18,10 @@ requireLeveldown();
 var errors = require('../../deps/errors');
 var merge = require('../../merge');
 var utils = require('../../utils');
+var isDeleted = require('../../deps/docs/is-deleted');
+var isLocalId = require('../../deps/docs/is-local-id');
+var processDocs = require('../../deps/docs/process-docs');
+var md5 = require('../../deps/md5');
 var migrate = require('../../deps/migrate');
 var Deque = require("double-ended-queue");
 
@@ -326,7 +330,7 @@ function LevelPouch(opts, callback) {
         return callback(errors.error(errors.MISSING_DOC, 'missing'));
       }
 
-      if (utils.isDeleted(metadata) && !opts.rev) {
+      if (isDeleted(metadata) && !opts.rev) {
         return callback(errors.error(errors.MISSING_DOC, "deleted"));
       }
 
@@ -402,7 +406,7 @@ function LevelPouch(opts, callback) {
     // parse the docs and give each a sequence number
     var userDocs = req.docs;
     var docInfos = userDocs.map(function (doc, i) {
-      if (doc._id && utils.isLocalId(doc._id)) {
+      if (doc._id && isLocalId(doc._id)) {
         return doc;
       }
       var newDoc = utils.parseDoc(doc, newEdits);
@@ -477,7 +481,7 @@ function LevelPouch(opts, callback) {
       }
 
       userDocs.forEach(function (doc) {
-        if (doc._id && utils.isLocalId(doc._id)) {
+        if (doc._id && isLocalId(doc._id)) {
           // skip local docs
           return checkDone();
         }
@@ -564,7 +568,7 @@ function LevelPouch(opts, callback) {
 
       function onLoadEnd(doc, key, attachmentSaved) {
         return function (data) {
-          utils.MD5(data).then(
+          md5(data).then(
             onMD5Load(doc, key, data, attachmentSaved)
           );
         };
@@ -597,7 +601,7 @@ function LevelPouch(opts, callback) {
             onLoadEnd(docInfo, key, attachmentSaved));
           continue;
         }
-        utils.MD5(data).then(
+        md5(data).then(
           onMD5Load(docInfo, key, data, attachmentSaved)
         );
       }
@@ -775,8 +779,8 @@ function LevelPouch(opts, callback) {
         if (err) {
           return callback(err);
         }
-        utils.processDocs(docInfos, api, fetchedDocs,
-          txn, results, writeDoc, opts, finish);
+        processDocs(docInfos, api, fetchedDocs, txn, results, writeDoc,
+          opts, finish);
       });
     });
   });
@@ -824,7 +828,7 @@ function LevelPouch(opts, callback) {
       var docstream = stores.docStore.readStream(readstreamOpts);
 
       var throughStream = through(function (entry, _, next) {
-        if (!utils.isDeleted(entry.value)) {
+        if (!isDeleted(entry.value)) {
           if (skip-- > 0) {
             next();
             return;
@@ -860,7 +864,7 @@ function LevelPouch(opts, callback) {
           }
           if (opts.inclusive_end === false && metadata.id === opts.endkey) {
             return next();
-          } else if (utils.isDeleted(metadata)) {
+          } else if (isDeleted(metadata)) {
             if (opts.deleted === 'ok') {
               doc.value.deleted = true;
               doc.doc = null;
@@ -1042,7 +1046,7 @@ function LevelPouch(opts, callback) {
       // metadata not cached, have to go fetch it
       stores.docStore.get(doc._id, function (err, metadata) {
         if (opts.cancelled || opts.done || db.isClosed() ||
-          utils.isLocalId(metadata.id)) {
+          isLocalId(metadata.id)) {
           return next();
         }
         docIdsToMetadata.set(doc._id, metadata);
