+++ lib/adapters/websql/websql-bulk-docs.js
@@ -2,6 +2,10 @@
 
 var utils = require('../../utils');
 var errors = require('../../deps/errors');
+var preprocessAttachments =
+  require('../../deps/docs/preprocess-attachments');
+var isLocalId = require('../../deps/docs/is-local-id');
+var processDocs = require('../../deps/docs/process-docs');
 
 var websqlUtils = require('./websql-utils');
 var websqlConstants = require('./websql-constants');
@@ -22,7 +26,7 @@ function websqlBulkDocs(req, opts, api, db, Changes, callback) {
 
   // Parse the docs, give them a sequence number for the result
   var docInfos = userDocs.map(function (doc) {
-    if (doc._id && utils.isLocalId(doc._id)) {
+    if (doc._id && isLocalId(doc._id)) {
       return doc;
     }
     var newDoc = utils.parseDoc(doc, newEdits);
@@ -248,9 +252,8 @@ function websqlBulkDocs(req, opts, api, db, Changes, callback) {
     }
   }
 
-  function processDocs() {
-    utils.processDocs(docInfos, api, fetchedDocs,
-      tx, results, writeDoc, opts);
+  function websqlProcessDocs() {
+    processDocs(docInfos, api, fetchedDocs, tx, results, writeDoc, opts);
   }
 
   function fetchExistingDocs(callback) {
@@ -267,7 +270,7 @@ function websqlBulkDocs(req, opts, api, db, Changes, callback) {
     }
 
     docInfos.forEach(function (docInfo) {
-      if (docInfo._id && utils.isLocalId(docInfo._id)) {
+      if (docInfo._id && isLocalId(docInfo._id)) {
         return checkDone(); // skip local docs
       }
       var id = docInfo.metadata.id;
@@ -303,7 +306,7 @@ function websqlBulkDocs(req, opts, api, db, Changes, callback) {
     });
   }
 
-  utils.preprocessAttachments(docInfos, 'binary', function (err) {
+  preprocessAttachments(docInfos, 'binary', function (err) {
     if (err) {
       return callback(err);
     }
@@ -313,7 +316,7 @@ function websqlBulkDocs(req, opts, api, db, Changes, callback) {
         if (err) {
           preconditionErrored = err;
         } else {
-          fetchExistingDocs(processDocs);
+          fetchExistingDocs(websqlProcessDocs);
         }
       });
     }, unknownError(callback), complete);
