+++ lib/adapters/websql/websql.js
@@ -2,6 +2,8 @@
 
 var utils = require('../../utils');
 var merge = require('../../merge');
+var isDeleted = require('../../deps/docs/is-deleted');
+var isLocalId = require('../../deps/docs/is-local-id');
 var errors = require('../../deps/errors');
 var parseHexString = require('../../deps/parse-hex');
 var binStringToBlob = require('../../deps/binary/binaryStringToBlobOrBuffer');
@@ -154,10 +156,10 @@ function WebSqlPouch(opts, callback) {
             var item = result.rows.item(i);
             var seq = item.seq;
             var metadata = JSON.parse(item.metadata);
-            if (utils.isDeleted(metadata)) {
+            if (isDeleted(metadata)) {
               deleted.push(seq);
             }
-            if (utils.isLocalId(metadata.id)) {
+            if (isLocalId(metadata.id)) {
               local.push(metadata.id);
             }
           }
