+++ lib/changes.js
@@ -1,5 +1,6 @@
 'use strict';
 var utils = require('./utils');
+var isDeleted = require('./deps/docs/is-deleted');
 var merge = require('./merge');
 var errors = require('./deps/errors');
 var EE = require('events').EventEmitter;
@@ -113,7 +114,7 @@ function processChange(doc, metadata, opts) {
     doc: doc
   };
 
-  if (utils.isDeleted(metadata, doc._rev)) {
+  if (isDeleted(metadata, doc._rev)) {
     change.deleted = true;
   }
   if (opts.conflicts) {
