+++ lib/deps/docs/is-local-id.js
@@ -1,0 +1,7 @@
+'use strict';
+
+function isLocalId(id) {
+  return (/^_local/).test(id);
+}
+
+module.exports = isLocalId;
