+++ lib/deps/docs/parse-doc.js
@@ -1,7 +1,7 @@
 'use strict';
 
-var errors = require('./errors');
-var uuid = require('./uuid');
+var errors = require('./../errors');
+var uuid = require('./../uuid');
 
 function toObject(array) {
   return array.reduce(function (obj, item) {
