+++ lib/deps/md5.js
@@ -1,5 +1,6 @@
 'use strict';
 
+var toPromise = require('./to-promise');
 var base64 = require('./binary/base64');
 var crypto = require('crypto');
 var Md5 = require('spark-md5');
@@ -41,7 +42,7 @@ function appendString(buffer, data, start, end) {
   buffer.appendBinary(data);
 }
 
-module.exports = function (data, callback) {
+module.exports = toPromise(function (data, callback) {
   if (!process.browser) {
     var base64 = crypto.createHash('md5').update(data).digest('base64');
     callback(null, base64);
@@ -72,4 +73,4 @@ module.exports = function (data, callback) {
     }
   }
   loadNextChunk();
-};
+});
