+++ lib/deps/migrate.js
@@ -2,7 +2,7 @@
 
 var fs = require('fs');
 var path = require('path');
-var utils = require('../utils');
+var isLocalId = require('./docs/is-local-id');
 var merge = require('../merge');
 var levelup = require('levelup');
 var through = require('through2').obj;
@@ -133,7 +133,7 @@ exports.localAndMetaStores = function (db, stores, callback) {
           startKey: '_',
           endKey: '_\xFF'
         }).pipe(through(function (ch, _, next) {
-          if (!utils.isLocalId(ch.key)) {
+          if (!isLocalId(ch.key)) {
             return next();
           }
           batches.push({
@@ -167,7 +167,7 @@ exports.localAndMetaStores = function (db, stores, callback) {
           }
           deletedSeqs[seq] = true;
           stores.bySeqStore.get(seq, function (err, resp) {
-            if (err || !utils.isLocalId(resp._id)) {
+            if (err || !isLocalId(resp._id)) {
               return next();
             }
             batches.push({
