+++ lib/replicate/gen-replication-id.js
@@ -1,6 +1,6 @@
 'use strict';
 
-var utils = require('../utils');
+var md5 = require('../deps/md5');
 
 // TODO: check CouchDB's replication id generation
 // Generate a unique id particular to this replication
@@ -10,12 +10,12 @@ function genReplicationId(src, target, opts) {
     return target.id().then(function (target_id) {
       var queryData = src_id + target_id + filterFun +
         JSON.stringify(opts.query_params) + opts.doc_ids;
-      return utils.MD5(queryData).then(function (md5) {
+      return md5(queryData).then(function (md5sum) {
         // can't use straight-up md5 alphabet, because
         // the char '/' is interpreted as being for attachments,
         // and + is also not url-safe
-        md5 = md5.replace(/\//g, '.').replace(/\+/g, '_');
-        return '_local/' + md5;
+        md5sum = md5sum.replace(/\//g, '.').replace(/\+/g, '_');
+        return '_local/' + md5sum;
       });
     });
   });
